infinite scroll, verified field in users table and minor fixed and modifs

+++ app/Livewire/LikePost.php
@@ -12,7 +12,11 @@ class LikePost extends Component
 
     public function mount($post)
     {
-        $this->isLiked = $post->checkLikedByUser(auth()->user());
+        if (!auth()->check()) {
+            $this->isLiked = false;
+        } else {
+            $this->isLiked = $post->checkLikedByUser(auth()->user());
+        }
         $this->likes = $post->likes->count();
     }
 
@@ -23,6 +27,10 @@ class LikePost extends Component
 
     public function like()
     {
+        if (!auth()->check()) {
+            return redirect()->route('login');
+        }
+        
         if($this->post->checkLikedByUser(auth()->user()))
         {
             $this->post->likes()->where('post_id', $this->post->id)->delete();
